requirements.txt
click==8.3.1
Flask==2.2.5
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.5
pillow==12.0.0
python-dateutil==2.8.2
PyWavelets==1.9.0
scipy==1.16.3
six==1.17.0
Werkzeug==3.1.4
